Refactored feed item loading states
