+++ packages/pwa/src/components/feedItems/FeedItemMarkdown.tsx
@@ -5,10 +5,9 @@ import type {FeedItem} from '@shared/types/feedItems.types';
 import {useFeedItemMarkdown} from '@sharedClient/services/feedItems.client';
 
 import {Text} from '@src/components/atoms/Text';
-import {ImportingFeedItem} from '@src/components/feedItems/ImportingFeedItem';
 import {Markdown} from '@src/components/Markdown';
 
-const ImportedFeedItemMarkdown: React.FC<{readonly feedItem: FeedItem}> = ({feedItem}) => {
+export const FeedItemMarkdown: React.FC<{readonly feedItem: FeedItem}> = ({feedItem}) => {
   const markdownState = useFeedItemMarkdown(feedItem);
 
   if (markdownState.error) {
@@ -28,14 +27,4 @@ const ImportedFeedItemMarkdown: React.FC<{readonly feedItem: FeedItem}> = ({feed
   }
 
   return <Text as="p">No markdown</Text>;
-};
-
-export const FeedItemMarkdown: React.FC<{readonly feedItem: FeedItem}> = ({feedItem}) => {
-  const hasFeedItemEverBeenImported = feedItem.importState.lastSuccessfulImportTime !== null;
-
-  if (hasFeedItemEverBeenImported) {
-    return <ImportedFeedItemMarkdown feedItem={feedItem} />;
-  }
-
-  return <ImportingFeedItem feedItem={feedItem} />;
 };
