+++ packages/pwa/src/components/feedItems/WebsiteFeedItem.tsx
@@ -2,18 +2,31 @@ import type React from 'react';
 
 import type {WebsiteFeedItem} from '@shared/types/feedItems.types';
 
-import {FeedItemHeader, FeedItemWrapper} from '@src/components/feedItems/FeedItem';
+import {FeedItemHeader, FeedItemSummary, FeedItemWrapper} from '@src/components/feedItems/FeedItem';
 import {FeedItemMarkdown} from '@src/components/feedItems/FeedItemMarkdown';
-import {FeedItemSummary} from '@src/components/feedItems/FeedItemSummary';
+import {ImportingFeedItem} from '@src/components/feedItems/ImportingFeedItem';
 
 export const WebsiteFeedItemComponent: React.FC<{readonly feedItem: WebsiteFeedItem}> = ({
   feedItem,
 }) => {
+  const hasFeedItemEverBeenImported = feedItem.importState.lastSuccessfulImportTime !== null;
+
+  let mainContent: React.ReactNode;
+  if (!hasFeedItemEverBeenImported) {
+    mainContent = <ImportingFeedItem feedItem={feedItem} />;
+  } else {
+    mainContent = (
+      <>
+        <FeedItemSummary feedItem={feedItem} />
+        <FeedItemMarkdown feedItem={feedItem} />
+      </>
+    );
+  }
+
   return (
     <FeedItemWrapper>
       <FeedItemHeader feedItem={feedItem} />
-      <FeedItemSummary feedItem={feedItem} />
-      <FeedItemMarkdown feedItem={feedItem} />
+      {mainContent}
     </FeedItemWrapper>
   );
 };
