+++ packages/pwa/src/components/feedItems/XkcdFeedItem.tsx
@@ -4,12 +4,22 @@ import type {XkcdFeedItem} from '@shared/types/feedItems.types';
 
 import {FeedItemHeader, FeedItemWrapper} from '@src/components/feedItems/FeedItem';
 import {FeedItemMarkdown} from '@src/components/feedItems/FeedItemMarkdown';
+import {ImportingFeedItem} from '@src/components/feedItems/ImportingFeedItem';
 
 export const XkcdFeedItemComponent: React.FC<{readonly feedItem: XkcdFeedItem}> = ({feedItem}) => {
+  const hasFeedItemEverBeenImported = feedItem.importState.lastSuccessfulImportTime !== null;
+
+  let mainContent: React.ReactNode;
+  if (!hasFeedItemEverBeenImported) {
+    mainContent = <ImportingFeedItem feedItem={feedItem} />;
+  } else {
+    mainContent = <FeedItemMarkdown feedItem={feedItem} />;
+  }
+
   return (
     <FeedItemWrapper>
       <FeedItemHeader feedItem={feedItem} />
-      <FeedItemMarkdown feedItem={feedItem} />
+      {mainContent}
     </FeedItemWrapper>
   );
 };
